Fix "debug" logging by only logging when a callback is present
If the client sets the log_header_callback, but not the
log_body_callback, debug would still be set to true. In the
post-response case, this would cause the XML to be pretty formatted for
no reason at all, which can be incredibly expensive. This new code only
does the message processing if the callback has been set.